made it installable
