+++ src/LeOpardLink.egg-info/requires.txt
@@ -1,0 +1,4 @@
+numpy
+pandas
+networkx
+jaal>=0.1.7
