+++ src/LeOpardLink/__init__.py
@@ -1,0 +1,8 @@
+"""
+LeOpardLink
+
+A package to generate all possible graphs for wildlife individual identification, and visualize them.
+"""
+
+__version__ = "0.1.0"
+__author__ = 'Courtney Allen, Ziqi Guo, Guy Bennevat Haninovich, Jiangyue Wang'
